lock down ci
whitelist PRs that are considered
require authorized users
