+++ ci2/ci/constants.py
@@ -1,3 +1,17 @@
 GITHUB_CLONE_URL = 'https://github.com/'
 # FIXME
 BUCKET = 'gs://user-buc4i59scmsjkhsm7iapev7es'
+
+AUTHORIZED_USERS = set([
+    'danking',
+    'cseed',
+    'konradjk',
+    'jigold',
+    'jbloom22',
+    'patrick-schultz',
+    'lfrancioli',
+    'akotlar',
+    'tpoterba',
+    'chrisvittal',
+    'catoverdrive'
+])
