+++ hailjwt/hailjwt/__init__.py
@@ -1,6 +1,7 @@
-from .hailjwt import JWTClient, get_domain
+from .hailjwt import JWTClient, get_domain, authenticated_users_only
 
 __all__ = [
-    JWTClient,
-    get_domain
+    'JWTClient',
+    'get_domain',
+    'authenticated_users_only'
 ]
